add function dhcpc_renew to make the process simple. add a check to the app_call so we only call the dhcp packet handler IF we have not aquired a dhcp lease. ... need to double check this for static ip...
Bring over change / fixed that the uhttpd-avr project found for the dhcp client
  THANK YOU AGAIN....
Fix the case where the device is getting LOTS of UDP packets while trying to aquire an IP via DHCP.
On my home network two of my home network routers and a few other devices have upnp and even with it turned off on these devices the network still gets a flood of SSDP packets every few seconds (5-8).flood being 50+ packets.

The change I made is to skip the timer adjust if we receive a packet.
The issue was that the we send out a Discover packet and while waiting for the Offer packet we get 50+ UDP packets which are NOT an Offer :-)  the timer is kicket up well over the 60 second limit and back around.  So, we totally miss the Offer packet.  With luck we get the process to work in a few hours of letting everything sit :-p

Anyhow, the change I made seems to fix the issue 95+% of my uses now.



git-svn-id: http://avr-uip.googlecode.com/svn/trunk@28 30fab446-694b-0410-acbb-ad9843a15c28