Fix: Issue 16: 	resolv does not free up old entries
If the resolv app fails to resolve a hostname then on the next call by the client 
they will be given a special ip of 0.0.0.1 to note the failure.

On the next call to lookup the same host name a new attempt at resolving the given
host name will be started.

Magic return values are bad but, I don't really want to require all caller to pass
a pointer to a int to check if the resolve worked.  So, I'm going with magic.



git-svn-id: http://avr-uip.googlecode.com/svn/trunk@170 30fab446-694b-0410-acbb-ad9843a15c28